refactor(database): migrate from SQLite to PostgreSQL and adjust schema for compatibility

+++ migrations/000005_create_versions_files_table.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE versions_files
 (
-    version_id INTEGER NOT NULL,
-    file_id    INTEGER NOT NULL,
+    version_id BIGINT NOT NULL,
+    file_id    BIGINT NOT NULL,
     PRIMARY KEY (version_id, file_id),
     CONSTRAINT fk_versions_files_version FOREIGN KEY (version_id) REFERENCES versions (id) ON DELETE CASCADE,
     CONSTRAINT fk_versions_files_file FOREIGN KEY (file_id) REFERENCES files (id)
